Refactor TokenFactory -> CookieExtractor
