+++ client-network/src/main/java/com/jaspersoft/android/sdk/network/api/AuthenticationRestApiImpl.java
@@ -73,7 +73,7 @@ final class AuthenticationRestApiImpl implements AuthenticationRestApi {
             com.squareup.okhttp.Response response = call.execute();
             int statusCode = response.code();
             if (statusCode >= 200 && statusCode < 300) { // 2XX == successful request
-                return TokenFactory.create(response);
+                return CookieExtractor.extract(response);
             } else if (statusCode >= 300 && statusCode < 400) { // 3XX == redirect request
                 String location = response.headers().get("Location");
                 if (location == null) {
@@ -82,7 +82,7 @@ final class AuthenticationRestApiImpl implements AuthenticationRestApi {
                 HttpUrl url = HttpUrl.parse(location);
                 String errorQueryParameter = url.queryParameter("error");
                 if (errorQueryParameter == null) {
-                    return TokenFactory.create(response);
+                    return CookieExtractor.extract(response);
                 } else {
                     com.squareup.okhttp.Response response401 = new com.squareup.okhttp.Response.Builder()
                             .protocol(response.protocol())
@@ -106,7 +106,7 @@ final class AuthenticationRestApiImpl implements AuthenticationRestApi {
     public EncryptionKey requestEncryptionMetadata() {
         RestApi api = mRestAdapterBuilder.build().create(RestApi.class);
         Response response = CallWrapper.wrap(api.requestAnonymousCookie()).response();
-        String anonymousToken = TokenFactory.create(response.raw());
+        String anonymousToken = CookieExtractor.extract(response.raw());
 
         RestApi modifiedApi = mRestAdapterBuilder.build().create(RestApi.class);
