+++ client-network/src/test/java/com/jaspersoft/android/sdk/network/api/CookieExtractorTest.java
@@ -24,43 +24,42 @@
 
 package com.jaspersoft.android.sdk.network.api;
 
+import com.squareup.okhttp.Protocol;
+import com.squareup.okhttp.Request;
 import com.squareup.okhttp.Response;
 
-import java.util.Iterator;
-import java.util.List;
+import org.junit.Before;
+import org.junit.Test;
 
+import static org.hamcrest.CoreMatchers.is;
+import static org.junit.Assert.assertThat;
 
 /**
  * @author Tom Koptel
  * @since 2.0
  */
-final class TokenFactory {
-    private final List<String> mCookieParts;
+public class CookieExtractorTest {
 
-    TokenFactory(List<String> parts) {
-        mCookieParts = parts;
+    private Request mRequest;
+
+    @Before
+    public void setup() {
+        mRequest = new Request.Builder()
+                .url("http://localhost")
+                .build();
     }
 
-    public static String create(Response response) {
-        List<String> parts = response.headers().values("Set-Cookie");
-        TokenFactory responseFactory = new TokenFactory(parts);
-        return responseFactory.create();
-    }
+    @Test
+    public void shouldExtractTokenFromNetworkResponse() {
+        Response mockResponse = new Response.Builder()
+                .addHeader("Set-Cookie", "cookie1")
+                .addHeader("Set-Cookie", "cookie2")
+                .code(200)
+                .protocol(Protocol.HTTP_1_1)
+                .request(mRequest)
+                .build();
 
-    private String create() {
-        return joinCookieParts().toString();
-    }
-
-    private StringBuilder joinCookieParts() {
-        StringBuilder stringBuilder = new StringBuilder();
-        Iterator<String> iterator = mCookieParts.iterator();
-        while (iterator.hasNext()) {
-            String cookie = iterator.next();
-            stringBuilder.append(cookie);
-            if (iterator.hasNext()) {
-                stringBuilder.append(";");
-            }
-        }
-        return stringBuilder;
+        String token = CookieExtractor.extract(mockResponse);
+        assertThat(token, is("cookie1;cookie2"));
     }
 }
